XMLParser.jl

A lightweight XML parser which is purely written in Julia.

For extensive use of XML features, consider using one of the following XML-libraries: LightXML.jl, EzXML.jl.

Install

import Pkg
Pkg.add("XMLParser")

Unsupported features

It is not allowed to use one of the following characters in XML content: <,>

There are probably many more XML features that are not supported at the moment.

XMLParser.jl

Install

Types

Functions

Types

XMLParser.XMLAttribute — Type

XMLAttribute

Creates a XML attribute key-value-pair

XMLAttribute(key::String, val::String)

XMLParser.XMLTag — Type

XMLTag

Creates a XML tag

XMLTag(name::String, attributes::Vector{XMLAttribute})

Missing docstring.

Missing docstring for XMLEmptyTag. Check Documenter's build log for details.

XMLParser.XMLElement — Type

XMLElement <: AbstractXMLElement

Creates a XML element

XMLElement(tag::XMLTag, content::Vector{Any})

Functions

Utils

XMLParser.getElements! — Function

getElements!(ret::Vector{XMLElement},xmlel::XMLElement,tagname::String)

Pushes all XMLElements that match the tagname to ret.

XMLParser.getElements — Function

getElements(xmlel::XMLElement,tagname::String)

Gets all XMLElements that match the tagname and returns a Vector{XMLElement}.

XMLParser.getChildrenbyTagName! — Function

getChildrenbyTagName!(ret::Vector{XMLElement},xmlel::XMLElement,tagname::String)

Pushes all XMLElements that are direct children of xmlel and match the tagname to ret.

XMLParser.getChildrenbyTagName — Function

getChildrenbyTagName(xmlel::XMLElement,tagname::String)

Gets all XMLElements that are direct children of xmlel and match the tagname. Return a Vector{XMLElement}.

XMLParser.hasAttributekey — Function

hasAttributekey(el::XMLElement,key::String))

Return true if el has a attribute with keyname key. Returns false otherwise.

XMLParser.getAttribute — Function

getAttribute(el::XMLElement,key::String)

Return attribute value if el has a attribute with keyname key. Returns emptystring otherwise.

XMLParser.setAttribute — Function

setAttribute(el::XMLElement,key::String,val)

Sets attribute value val to XML attribute key of XMLElement el. Returns an error("Attribute not found") if key is not found. Returns nothing otherwise.

IO-Functions

Base.read — Method

Base.read(::Type{XMLElement}, file::String)

Reads a XML file in location file. Returns a XMLElement.

XMLParser.writeXMLElement — Function

writeXMLElement(f::IOStream, el::XMLElement)

Writes a XMLElement to an IOStream.

writeXMLElement(f::IOStream, el::XMLElement)

Writes a XMLElement to an IOStream.